+++ .eslintrc.json
@@ -29,5 +29,8 @@
             "error",
             "always"
         ]
+	},
+	"parserOptions": {
+        "sourceType": "module"
     }
 }
