+++ app/js/play.js
@@ -1,3 +1,5 @@
+import { randomNoteGenerator } from './libs/randomNoteGenerator';
+
 window.onload = randomNoteGenerator();
 
 document.addEventListener('keydown', whichKey);
@@ -34,23 +36,6 @@ function removeClassOnTransitionEnd(e) {
 
 const keys = document.querySelectorAll('.key');
 keys.forEach(key => key.addEventListener('transitionend', removeClassOnTransitionEnd));
-
-function randomNoteGenerator() {
-	let possibleNotesArray = [
-		['a', '72'], ['a#', '85'], ['b', '74'], ['c', '65'], ['c#', '87'],
-		['d', '83'], ['d#', '69'], ['e', '68'], ['f', '70'], ['f#', '84'],
-		['g', '71'], ['g#', '89']
-	];
-	let randomNote = possibleNotesArray[Math.floor(Math.random() * possibleNotesArray.length)];
-	let lastRandomNote = document.getElementById('targetNote').dataset.key;
-	if (lastRandomNote == randomNote) {
-		let randomNote = possibleNotesArray[Math.floor(Math.random() * possibleNotesArray.length)];
-		return randomNote;
-	}
-	document.getElementById('targetNote').innerHTML= randomNote[0];
-	document.getElementById('targetNote').dataset.key = randomNote[1];
-	return randomNote;
-}
 
 function compareNoteIsCorrect(randomNote, pressedNote, audio, key) {
 	let audioFail = document.querySelector(`audio[data-key='x']`);
